+++ src/test/java/com/github/alexdlaird/ngrok/conf/JavaNgrokConfigTest.java
@@ -28,13 +28,19 @@ import com.github.alexdlaird.ngrok.process.NgrokLog;
 import com.github.alexdlaird.ngrok.protocol.Region;
 import org.junit.jupiter.api.Test;
 
+import java.lang.reflect.Field;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Map;
 import java.util.function.Function;
 
+import static com.github.alexdlaird.util.StringUtils.isBlank;
+import static com.github.alexdlaird.util.StringUtils.isNotBlank;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 public class JavaNgrokConfigTest {
     @Test
@@ -81,5 +87,19 @@ public class JavaNgrokConfigTest {
     public void testJavaNgrokConfigWithInvalidStartupTimeout() {
         // WHEN
         assertThrows(IllegalArgumentException.class, () -> new JavaNgrokConfig.Builder().withStartupTimeout(0));
+    }
+
+    @Test
+    public void testAuthTokenSetFromEnv() {
+        // GIVEN
+        final String ngrokAuthToken = System.getenv("NGROK_AUTHTOKEN");
+        assumeTrue(isNotBlank(System.getenv("NGROK_AUTHTOKEN")), "NGROK_AUTHTOKEN environment variable not set");
+
+        // WHEN
+        final JavaNgrokConfig javaNgrokConfig = new JavaNgrokConfig.Builder()
+                .build();
+
+        // THEN
+        assertEquals(ngrokAuthToken, javaNgrokConfig.getAuthToken());
     }
 }